Use Long types in client

+++ android/app/src/main/java/ru/tesseract/assets/domain/DetailedAssetInfo.kt
@@ -25,9 +25,9 @@ private class DetailedAssetInfoSurrogate(
     @SerialName("companyTitle")
     val companyTitle: String,
     @SerialName("assetPrice")
-    val price: Int,
+    val price: Long,
     @SerialName("assetPriceDiff")
-    val priceDiff: Int,
+    val priceDiff: Long,
     @SerialName("favouriteStatus")
     val isFavorite: Boolean,
     @SerialName("assetDescription")
